requirements.txt
sentence-transformers==2.7.0
mistralai==0.4.2
transformers==4.40.0
faiss-cpu==1.8.0
numpy==1.24.3
pandas==2.2.2
python-dotenv==1.0.1
tqdm==4.66.4
torch==2.3.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
